Update project4.ino
This code has DHT11 temperature and humidity sensors, an analog light sensor and a gas sensor. It assumes that you connect the pin of the first DHT11 sensor to pin 2, the pin of the second DHT11 sensor to pin 3, the pin of the analog sensor to pin A0, and the pin of the gas sensor to pin A1.

This code reads the data from each sensor separately and prints it to the serial monitor. The value from the gas sensor measures the gas level in the environment.